Move resolver related bits into resolver module
This change moves more functionality out of lib.rs. Specifically, we
move everything resolver related into the newly introduced resolver
module.
